refactor(mcp): split churned domains out of register-tools monolith
register-tools.ts was a 2467-line file every MCP change had to edit,
maximising merge-clash surface on the most active features. Peels the
newest/most-churned domains — plugins, industry presets, Public Form
Builder, Client Onboarding, Contracts (30 tools) — into
src/lib/mcp/tools/plugin-form-tools.ts, and the 4 helpers they share with
the main file (ctxFrom, appBase, UUID, getOrMintPortalToken) into
src/lib/mcp/tools/shared.ts.

Behaviour is identical: tools were moved verbatim and are registered
through the same tool() helper via registerPluginFormTools(tool). Tool
count unchanged (112 + 30 = 142). Monolith drops ~520 lines.

Verified: tsc clean, 17 tests pass, full build green, /api/mcp compiles.
